moved user logic to custom hook
